Split protego.py, add py.typed.
