+++ src/protego/__init__.py
@@ -1,0 +1,4 @@
+from ._protego import Protego
+from ._ruleset import RequestRate, VisitTime
+
+__all__ = ["Protego", "RequestRate", "VisitTime"]
